Rename file
This is where the open-source LCD code was pasted on top of the VHDL file from the TDC. 

The TDC begins on line 872, ends on line 1019.

This file was used to test how to initialize everything within one file. (note: integration was not saved on this, this is simply putting the two vhdl codes into one spot. Quartus compilation will fail for this file, as in.)