Rename file
This is a file which was entitled "TDCV1.vhd" 

This file was used to test how to initialize the TDC within the LCD code. It does not feature any of the new student code. It only features an attempt that was made to combine the TDC vhdl with the LCD vhdl code.